refactor AspNetCore test: ApplicationInsightsEXtensionsTests
