--- NETCORE/test/Microsoft.ApplicationInsights.AspNetCore.Tests/Properties/launchSettings.json
@@ -1,19 +1,0 @@
-{
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:4822/",
-      "sslPort": 0
-    }
-  },
-  "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNET_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
